predictive collision working, too much impact on performance, broken workflow with kinematica/character controller
it would be stupid to continue trying to use kinematica when every step I take forward creates a problem with it. The way it is now, the animations go first and the controller has to adapt, which does not make any sense if, for example, you want to detect collisions witrh the character controller's capsule. If I want to accomplish the objectives of this work, I need to get rid of kinematica and its dependencies. I have my own controller and abillities, I need to adapt to the typical animator and controller, and sample data from it