New Warper!!! applying custom root motion
tested with jog vault ledge mainly, need to adapt the other anims. But I've just killed the hiccups, there are no more teleports, transform follows skeleton where it goes, so there is no longer need for adjusting skeleton due to ik bug. This is an enormous breakthrough. I have to polish it, and adapt everything, check it warps properly over longer distances... But it is a great start!